Fixes #3
The old code stored the set of channels a player was subscribed to as
a `Set<String>` in `EntityPlayerMP`. Unfortunately, when a player dies
or returns to the overworld from the end, their `EntityPlayerMP`
instance is destroyed and a new one is created, causing the set of
registered channels to be lost. This commit tracks player registrations
in a new singleton class `PluginChannelTracker`.